requirements.txt
selenium
requests
colorama
flask
pyngrok
Pillow
